# portal_iyte_be/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Flask project",
        "version": "1.0.0",
        "title": "Flask API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "contact@yourapi.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:5000",
    "basePath": "/api",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "prefix + '/post/<query>": {
            "get": {
                "summary": "Add a brief description",
                "description": "Detailed description of the endpoint",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "prefix + '/user/<query>": {
            "get": {
                "summary": "Add a brief description",
                "description": "Detailed description of the endpoint",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "prefix + \"/user/<user_id>": {
            "get": {
                "summary": "Add a brief description",
                "description": "Detailed description of the endpoint",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "prefix + '/post": {
            "post": {
                "summary": "Add a brief description",
                "description": "Detailed description of the endpoint",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
...
}